requirements.txt
biopython
pyyaml
numpy
svgwrite
bcbio-gff
pypng
pygal
